# [CPP] Add bounds checking