Changed all static arrays with std::array objects. Also changed all [i] -> at(i) (to use bounds checking)

In this line 'relatedPost = {p._id, p.tags, {0,0,0,0,0}};' I've added explicit initialization with zeros. When I tried to put there 'std::array<Posts const*, TOPN>' - it showed me error.
Initially there were used std::vector{TOPN}.
But it seems vector is more suitable for dynamic arrays, while std::array for fixed sized (https://en.cppreference.com/w/cpp/container/array).

I've tested only on M1 clang++ : the performance became slower. Maybe with g++ it will be better.

Review and suggestions from C++-guys will be appreciated @kaiwu @jinyus @Moelf