@Moelf 2023-11-01
I see 17,18 -> 19 ms on my machine which isn't too bad honestly