@Moelf 2023-11-01
FIY I tried to apply these changes to `cpp_con` and I get out of bound error at runtime ...