@jinyus 2023-11-01
Hardcoded initialization `{0,0,0,0,0}` isn't allowed so I will mark this as draft until someone more familiar with the lang fixes it. I will try LLMs