@jinyus 2023-11-01
Much slower, would like to see a version with just the `.at` changes to compare.

Main:
`C++ | 24.92 ms | 339.00 ms | 2.92 s`

This PR:
```
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 38 ms
        Processing time (w/o IO): 37 ms
        Processing time (w/o IO): 37 ms
        Processing time (w/o IO): 38 ms
        Processing time (w/o IO): 37 ms
        Processing time (w/o IO): 37 ms
        Processing time (w/o IO): 38 ms
        Processing time (w/o IO): 38 ms
        Processing time (w/o IO): 37 ms
        Processing time (w/o IO): 37 ms
        Processing time (w/o IO): 37 ms
        Processing time (w/o IO): 37 ms
        Processing time (w/o IO): 37 ms
          Time (mean ± σ):     235.3 ms ±   2.0 ms    [User: 207.9 ms, System: 27.1 ms]
          Range (min … max):   232.5 ms … 239.0 ms    10 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 555 ms
        Processing time (w/o IO): 562 ms
        Processing time (w/o IO): 558 ms
          Time (mean ± σ):      1.377 s ±  0.002 s    [User: 1.249 s, System: 0.121 s]
          Range (min … max):    1.375 s …  1.378 s    2 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 4918 ms
        Processing time (w/o IO): 4917 ms
        Processing time (w/o IO): 4909 ms
          Time (mean ± σ):      7.551 s ±  0.068 s    [User: 7.027 s, System: 0.346 s]
          Range (min … max):    7.503 s …  7.600 s    2 runs
```