@Moelf 2023-11-01
>Much slower

FWIW, that's largely real because I see a 20% speed-up in Julia by doing `inbounds`