@cyrusmsk 2023-11-01
> Much slower, would like to see a version with just the `.at` changes to compare.

but this is the version with just “at”. We can return back the initialization to std::vector. But at my machine there is no any changes in that