@Moelf 2023-11-01
>It's just weird because locally, it's about 2ms slower.

yeah we all know VM is weird at this point haha