@Moelf 2023-11-01
VM is nice at least it's reproducible (not everyone has the same Intel/AMD/Apple silicon) but in this case because it's vCPU I feel like somehow it's not reproducible in a meaningful way